backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
python-dotenv>=1.0.0
openai>=1.0.0